Taken out service abstraction
