+++ ui/src/app.rs
@@ -1,6 +1,7 @@
 use super::components;
 use super::services;
-use services::RbpService;
+use services::rpb_service::RbpService;
+use yew::macros::html;
 use yew::prelude::*;
 use yew::virtual_dom::VNode;
 use yew_router::{prelude::*, Switch};
